moved asset movement function to separate file
